+++ app/src/main/AndroidManifest.xml
@@ -2,8 +2,8 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     package="org.sopt.sample">
-    <uses-permission android:name="android.permission.INTERNET" />
 
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
         android:allowBackup="true"
@@ -34,7 +34,7 @@
         </activity>
 
         <activity
-            android:name="com.sopt.hajeong.main.MainActivity"
+            android:name="com.sopt.hajeong.main.view.MainActivity"
             android:exported="false">
 
         </activity>
