+++ app/src/main/java/com/sopt/hajeong/data/api/ApiFactory.kt
@@ -1,24 +1,23 @@
-package com.sopt.hajeong.data
+package com.sopt.hajeong.data.api
 
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
 import retrofit2.Retrofit
-import retrofit2.create
 
 object ApiFactory {
-    private val soptRetrofit: Retrofit by lazy {
+    val authRetrofit: Retrofit by lazy {
         Retrofit.Builder()
             .baseUrl("http://3.39.169.52:3000/")
             .addConverterFactory(Json.asConverterFactory("application/json".toMediaType())) //Json데이터를 사용자가 정의한 Java 객채로 변환해주는 라이브러리
             .build() //레트로핏 구현체 완성!
     }
-    private val reqresRetrofit: Retrofit by lazy {
+    val reqresRetrofit: Retrofit by lazy {
         Retrofit.Builder()
             .baseUrl("https://reqres.in/")
             .addConverterFactory(Json.asConverterFactory("application/json".toMediaType())) //Json데이터를 사용자가 정의한 Java 객채로 변환해주는 라이브러리
             .build() //레트로핏 구현체 완성!
     }
-    val authService: AuthService= soptRetrofit.create(AuthService::class.java)
-    val followerService: FollowerService= reqresRetrofit.create(FollowerService::class.java)
+    val authService: AuthService = authRetrofit.create(AuthService::class.java)
+    val followerService: FollowerService = reqresRetrofit.create(FollowerService::class.java)
 }
