+++ app/src/main/java/com/sopt/hajeong/data/model/ResponseGetFollowerListDTO.kt
