+++ app/src/main/java/com/sopt/hajeong/main/adapter/FollowerAdapter.kt
@@ -1,18 +1,18 @@
-package com.sopt.hajeong.main
+package com.sopt.hajeong.main.adapter
 
 import android.content.Context
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.sopt.hajeong.data.ResponseGetFollowerListDTO
+import com.sopt.hajeong.data.model.ResponseGetFollowerListDTO
 import org.sopt.sample.databinding.ItemHomeBodyBinding
 
 class FollowerAdapter(context: Context) : RecyclerView.Adapter<FollowerAdapter.BodyViewHolder>() {
     private val inflater by lazy { LayoutInflater.from(context) }
     private var followerList: List<ResponseGetFollowerListDTO.Follower> = emptyList()
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int):BodyViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): BodyViewHolder {
         val binding = ItemHomeBodyBinding.inflate(inflater, parent, false)
         return BodyViewHolder(binding)
     }
@@ -34,10 +34,9 @@ class FollowerAdapter(context: Context) : RecyclerView.Adapter<FollowerAdapter.B
         private val binding: ItemHomeBodyBinding
     ) : RecyclerView.ViewHolder(binding.root) {
         fun onBind(follower: ResponseGetFollowerListDTO.Follower) {
-            //Glide 라이브러리: 사진 로딩에 특화됨. 메모리 절약과 자연스러운 사진 로딩에 사용
             Glide.with(this.binding.root)
-                .load(follower.avatar)
-                .into(binding.ivProfile)
+                .load(follower.avatar)//여기서 url을 받아와야지
+                .into(binding.ivProfile)//여기서 뷰바인딩 해야지
             binding.tvName.text=follower.firstName
             binding.tvEmail.text=follower.email
         }
